# Conflicts:
#	server/game/DLLoader.hpp

+++ client/include/entities/PlayerAndStarshipEntity.hpp
@@ -26,13 +26,13 @@ class PlayerAndStarshipEntity : public Engine::Entity {
 
         this->addComponent<Engine::TransformComponent>();
         auto starship = std::make_unique<SpriteSFML>(starshipTexture);
-        starship->setScale({static_cast<float>(STARSHIP_SCALE_X), static_cast<float>(STARSHIP_SCALE_Y)});
-        starship->setRect({Engine::Box<int>({STARSHIP_WIDTH * 4, 0}, {STARSHIP_WIDTH, STARSHIP_HEIGHT})});
-        starship->setOrigin(Engine::Point<float>{-200, 0});
+        starship->setScale({static_cast<double>(STARSHIP_SCALE_X), static_cast<double>(STARSHIP_SCALE_Y)});
+        starship->setRect({Engine::Box<double>({STARSHIP_WIDTH * 4, 0}, {STARSHIP_WIDTH, STARSHIP_HEIGHT})});
+        starship->setOrigin(Engine::Point<double>{-200, 0});
         this->addComponent<Engine::SpriteComponent>(2, std::move(starship));
 
         auto playerNameText = std::make_unique<TextSFML>(playerName, font, 80);
-        playerNameText->setOrigin(Engine::Point<float>{0, 40.0});
+        playerNameText->setOrigin(Engine::Point<double>{0, 40.0});
         this->addComponent<Engine::TextComponent>(2, std::move(playerNameText));
     }
 };
