+++ client/include/sfml/TextSFML.hpp
@@ -30,20 +30,20 @@ class TextSFML : public Engine::AText {
             setCharacterSize(characterSize);
         }
         void setFont(std::shared_ptr<Engine::AFont> &font) final;
-        void setScale(Engine::Size<float> scale) final;
-        void setOrigin(Engine::Point<float> origin) final;
+        void setScale(Engine::Size<double> scale) final;
+        void setOrigin(Engine::Point<double> origin) final;
         void setFillColor(Engine::Color color) final;
         void setOutlineColor(Engine::Color color) final;
         void setString(std::string text) final;
-        void draw(std::shared_ptr<Engine::AWindow> &window, Engine::Point<int> position, float angle) final;
-        void setRotation(float angle) final;
-        void setPosition(Engine::Point<float> position) final;
+        void draw(std::shared_ptr<Engine::AWindow> &window, Engine::Point<double> position, double angle) final;
+        void setRotation(double angle) final;
+        void setPosition(Engine::Point<double> position) final;
         void setCharacterSize(unsigned int size) final;
-        void setLetterSpacing(float spacingFactor) final;
-        void setLineSpacing(float spacingFactor) final;
-        [[nodiscard]] Engine::Point<float> getPosition() const final;
+        void setLetterSpacing(double spacingFactor) final;
+        void setLineSpacing(double spacingFactor) final;
+        [[nodiscard]] Engine::Point<double> getPosition() const final;
         [[nodiscard]] Engine::Size<int> getSize() const final;
-        [[nodiscard]] float getRotation() const final;
+        [[nodiscard]] double getRotation() const final;
         [[nodiscard]] std::string toStdString() const final;
 };
